rename dependency to provider
