+++ morphine/__init__.py
@@ -1,2 +1,2 @@
 from morphine.domain.container import Container
-from morphine.domain.dependency import Dependency
+from morphine.domain.provider import Provider
