+++ pyproject.toml
@@ -15,6 +15,7 @@ python = "^3.11"
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.0.0"
 black = "^24.1.1"
+docformatter = "^1.7.5"
 
 
 [tool.poetry.group.tests.dependencies]
